backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic[email]==2.9.2
PyJWT==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
